+++ src/PagesInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace iio\libmergepdf;
+
+interface PagesInterface
+{
+    /**
+     * @return int[]
+     */
+    public function getPageNumbers(): array;
+}