+++ tests/PagesTest.php
@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace iio\libmergepdf;
 
-class PagesTest extends \PHPUnit_Framework_TestCase
+class PagesTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @dataProvider pageNumbersProvider
@@ -11,7 +13,7 @@ class PagesTest extends \PHPUnit_Framework_TestCase
     {
         $this->assertSame(
             $expected,
-            iterator_to_array(new Pages($expressionString))
+            (new Pages($expressionString))->getPageNumbers()
         );
     }
 
@@ -28,25 +30,9 @@ class PagesTest extends \PHPUnit_Framework_TestCase
         ];
     }
 
-    /**
-     * @expectedException iio\libmergepdf\Exception
-     */
     public function testInvalidString()
     {
+        $this->expectException(Exception::CLASS);
         new Pages('12,*');
-    }
-
-    public function testIsIterabla()
-    {
-        $this->assertSame(
-            [1, 2],
-            iterator_to_array(new Pages('1, 2'))
-        );
-    }
-
-    public function testHasPages()
-    {
-        $this->assertFalse((new Pages)->hasPages());
-        $this->assertTrue((new Pages('1'))->hasPages());
     }
 }